+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * add `destroy-database!` to core namespace
 * deprecate `open-database` and replace it with `open-database!`
 * deprecate `close-database` and replace it with `close-database!`
-* [TODO] `close-all-databases!` function - closes all open databases
+* `close-all-databases!` function - closes all open databases
 * `is-open?` database predicate
 
 ## 1.0.3 (2017-10-12)